Incorporating new changes - sign up with self-defined username

+++ Domain/Models/User.cs
@@ -8,7 +8,7 @@ namespace Domain.Models
 {
     public class User
     {
-        public ulong Id { get; set; }
+        public int Id { get; set; }
         public string FullName { get; set; }
         public string Email { get; set; }
         public string UserName { get; set; }
